+++ source/libamtpa/client_access/message/amtp_upload_file_data_resp.cpp
@@ -25,7 +25,7 @@ void amtp_upload_file_data_resp::data(JwumqMessage *msg, UPLOAD_FILE_DATA_RESP_S
 	cmdprimitive.ParseFromArray(msg->RawData(),msg->RawDataLen());
 	
 	amtpap::UploadFileDataRespV1 resp;
-	resp.ParseFromArray(cmdprimitive.payload_data().c_str(), cmdprimitive.payload_data().size());
+	resp.ParseFromArray(cmdprimitive.payload_data().c_str(), (int)cmdprimitive.payload_data().size());
 	
 	// UPLOAD_FILE_DATA_RESP_STRU s;
 	memset(&s, 0, sizeof(UPLOAD_FILE_DATA_RESP_STRU));
